Remove unneeded casting

+++ src/main/java/net/danygames2014/nyatec/world/feature/OreFeature.java
@@ -28,40 +28,38 @@ public class OreFeature extends Feature {
         // Low Value = High Cosine = Wide Z Spread
         float spreadShape = random.nextFloat() * (float) Math.PI;
 
-        double spreadX1 = (float) (x + 8) + ((MathHelper.sin(spreadShape) * (float) this.oreCount) / 8.0F);
-        double spreadX2 = (float) (x + 8) - ((MathHelper.sin(spreadShape) * (float) this.oreCount) / 8.0F);
-        double spreadZ1 = (float) (z + 8) + ((MathHelper.cos(spreadShape) * (float) this.oreCount) / 8.0F);
-        double spreadZ2 = (float) (z + 8) - ((MathHelper.cos(spreadShape) * (float) this.oreCount) / 8.0F);
+        double spreadX1 = (x + 8) + ((MathHelper.sin(spreadShape) * this.oreCount) / 8.0D);
+        double spreadX2 = (x + 8) - ((MathHelper.sin(spreadShape) * this.oreCount) / 8.0D);
+        double spreadZ1 = (z + 8) + ((MathHelper.cos(spreadShape) * this.oreCount) / 8.0D);
+        double spreadZ2 = (z + 8) - ((MathHelper.cos(spreadShape) * this.oreCount) / 8.0D);
         double spreadY1 = y + random.nextInt(3) + 2;
         double spreadY2 = y + random.nextInt(3) + 2;
 
         for (int oreIndex = 0; oreIndex <= this.oreCount; ++oreIndex) {
-            double oreOffsetX = spreadX1 + (((spreadX2 - spreadX1) * (double) oreIndex) / (double) this.oreCount);
-            double oreOffsetY = spreadY1 + (spreadY2 - spreadY1) * (double) oreIndex / (double) this.oreCount;
-            double oreOffsetZ = spreadZ1 + (spreadZ2 - spreadZ1) * (double) oreIndex / (double) this.oreCount;
-            double veinSizeModifier = random.nextDouble() * (double) this.oreCount / (double) 16.0F;
-            double halfWidth = (double) (MathHelper.sin((float) oreIndex * (float) Math.PI / (float) this.oreCount) + 1.0F) * veinSizeModifier + (double) 1.0F;
-            double halfHeight = (double) (MathHelper.sin((float) oreIndex * (float) Math.PI / (float) this.oreCount) + 1.0F) * veinSizeModifier + (double) 1.0F;
-            int startX = MathHelper.floor(oreOffsetX - halfWidth / (double) 2.0F);
-            int startY = MathHelper.floor(oreOffsetY - halfHeight / (double) 2.0F);
-            int startZ = MathHelper.floor(oreOffsetZ - halfWidth / (double) 2.0F);
-            int endX = MathHelper.floor(oreOffsetX + halfWidth / (double) 2.0F);
-            int endY = MathHelper.floor(oreOffsetY + halfHeight / (double) 2.0F);
-            int endZ = MathHelper.floor(oreOffsetZ + halfWidth / (double) 2.0F);
+            double oreOffsetX = spreadX1 + (((spreadX2 - spreadX1) * oreIndex) / this.oreCount);
+            double oreOffsetY = spreadY1 + (spreadY2 - spreadY1) * oreIndex / this.oreCount;
+            double oreOffsetZ = spreadZ1 + (spreadZ2 - spreadZ1) * oreIndex / this.oreCount;
+            double veinSizeModifier = random.nextDouble() * this.oreCount / 16.0D;
+             double halfWidth = (MathHelper.sin(oreIndex * (float) Math.PI / this.oreCount) + 1.0D) * veinSizeModifier + 1.0D;
+            double halfHeight = (MathHelper.sin(oreIndex * (float) Math.PI / this.oreCount) + 1.0D) * veinSizeModifier + 1.0D;
+            int startX = MathHelper.floor(oreOffsetX - halfWidth / 2.0D);
+            int startY = MathHelper.floor(oreOffsetY - halfHeight / 2.0D);
+            int startZ = MathHelper.floor(oreOffsetZ - halfWidth / 2.0D);
+            int endX = MathHelper.floor(oreOffsetX + halfWidth / 2.0D);
+            int endY = MathHelper.floor(oreOffsetY + halfHeight / 2.0D);
+            int endZ = MathHelper.floor(oreOffsetZ + halfWidth / 2.0D);
 
             for (int oreX = startX; oreX <= endX; ++oreX) {
-                double wtfX = ((double) oreX + (double) 0.5F - oreOffsetX) / (halfWidth / (double) 2.0F);
-                if (wtfX * wtfX < (double) 1.0F) {
-                    
+                double wtfX = (oreX + 0.5D - oreOffsetX) / (halfWidth / 2.0D);
+                if (wtfX * wtfX < 1.0D) {
                     for (int oreY = startY; oreY <= endY; ++oreY) {
-                        double wtfY = ((double) oreY + (double) 0.5F - oreOffsetY) / (halfHeight / (double) 2.0F);
-                        if (wtfX * wtfX + wtfY * wtfY < (double) 1.0F) {
-                            
+                        double wtfY = (oreY + 0.5D - oreOffsetY) / (halfHeight / 2.0D);
+                        if (wtfX * wtfX + wtfY * wtfY < 1.0D) {
                             for (int oreZ = startZ; oreZ <= endZ; ++oreZ) {
-                                double wtfZ = ((double) oreZ + (double) 0.5F - oreOffsetZ) / (halfWidth / (double) 2.0F);
-                                if (wtfX * wtfX + wtfY * wtfY + wtfZ * wtfZ < (double) 1.0F) {
+                                double wtfZ = (oreZ + 0.5D - oreOffsetZ) / (halfWidth / 2.0D);
+                                if (wtfX * wtfX + wtfY * wtfY + wtfZ * wtfZ < 1.0D) {
                                     placeOre(world, oreX, oreY, oreZ);
-                                }
+                                } 
                             }
                         }
                     }
